__init__.py handling
