+++ sfs/time/wfs.py
@@ -44,11 +44,9 @@
 """
 import numpy as np
 from numpy.core.umath_tests import inner1d  # element-wise inner product
-from scipy.signal import besselap, sosfilt, zpk2sos
-from scipy.special import eval_legendre as legendre
 from .. import default
 from .. import util
-from . import source as _source
+from . import secondary_source_point, apply_delays
 
 
 def plane_25d(x0, n0, n=[0, 1, 0], xref=[0, 0, 0], c=None):
@@ -318,45 +316,3 @@ def driving_signals(delays, weights, signal):
     weights = util.asarray_1d(weights)
     data, samplerate, signal_offset = apply_delays(signal, delays)
     return util.DelayedSignal(data * weights, samplerate, signal_offset)
-
-
-def apply_delays(signal, delays):
-    """Apply delays for every channel.
-
-    Parameters
-    ----------
-    signal : (N,) array_like + float
-        Excitation signal consisting of (mono) audio data and a sampling
-        rate (in Hertz).  A `DelayedSignal` object can also be used.
-    delays : (C,) array_like
-        Delay in seconds for each channel (C), negative values allowed.
-
-    Returns
-    -------
-    `DelayedSignal`
-        A tuple containing the delayed signals (in a `numpy.ndarray`
-        with shape ``(N, C)``), followed by the sampling rate (in Hertz)
-        and a (possibly negative) time offset (in seconds).
-
-    """
-    data, samplerate, initial_offset = util.as_delayed_signal(signal)
-    data = util.asarray_1d(data)
-    delays = util.asarray_1d(delays)
-    delays += initial_offset
-
-    delays_samples = np.rint(samplerate * delays).astype(int)
-    offset_samples = delays_samples.min()
-    delays_samples -= offset_samples
-    out = np.zeros((delays_samples.max() + len(data), len(delays_samples)))
-    for column, row in enumerate(delays_samples):
-        out[row:row + len(data), column] = data
-    return util.DelayedSignal(out, samplerate, offset_samples / samplerate)
-
-
-def secondary_source_point(c):
-    """Create a point source for use in `sfs.time.synthesize()`."""
-
-    def secondary_source(position, _, signal, observation_time, grid):
-        return _source.point(position, signal, observation_time, grid, c=c)
-
-    return secondary_source
